Add Initial LDAP Configuration
Prior to this commit, there was no mechanism by which a live LDAP test
could be conducted.  This commit introduces a dependency on `exldap` for
the purposes of testing the Ecto LDAP Adapter.  Please note, user
credentials must be provided in the `dev.exs` configuration file.

Additionally, this commit provides a script to initialize the IEx shell
for the purposes of testing the adapter.  The README.md has also been
updated to track outstanding work and to remove default project text.