Create Ldap Adapter Registered Process
Prior to this commit, we weren't actually connecting to LDAP with our
search terms. This commit add a GenServer client API which is used by
the Ecto Adapter API to retrieve the information from LDAP. Eventually
we'll want to pull the GenServer client and server API portions out into
wherever the connection pool will live, but for now, this is fine and
semi-functional. Ecto.Ldap.Adapter.execute still fails; however, for an
entirely new reason, which is that our LDAP search results do not
conform to whatever contract `execute` expects.