Rename test script for automatic IEx usage
Prior to this commit, the `chuck.exs` script had to be manually imported
to an IEx session in order to avoid re-typing several `requires` and
`alias` statements.

With this commit, the file has been renamed from `chuck.exs` to
`.iex.exs`.  By doing this, running IEx will automatically source-in the
contents of `.iex.exs` without needing to enter a command at the IEx
prompt to import the file.  This will improve our speed with respect to
IEx debugging and development.

Please note: This script should be altered or removed once active
testing and development has completed.