Improve `execute` response data
Prior to this commit, raw data containing information from LDAP was
being returned as part of the `execute` method in `Ecto.Ldap.Adapter`.
Although the data was structured, it was not mapped into the appropriate
Ecto schema type as defined by the repo consumer.

This commit interprets the response data from LDAP and invokes the
appropriate mapping mechanisms to create an instance of the requested
model.  With this change, consumers of the LDAP-backed repository will
now receive back a list of valid model objects.

Please note, this is a work in progress and several features are still
missing, including the use of the `select:` option.