Clean up testing files
Prior to this commit, test files were colocated with production source
files in the `lib/ecto_ldap` directory.  Although useful for development
purposes, it breaks standard conventions to keep test files alongside
production source code.

This commit moves the files to the appropriate support location and
updates the `.iex.exs` script to reference the new location.
`test_model.ex` is also being removed as it is not currently being used.