break .get in preference for .get_by and all/2
Prior to this commit, we had a working get/2, but it relied on hacking
around the search terms after they'd already been translated. This
commit, rolls that functionality back for more general incremental
improvement of populating the filters with parameters and options, which
enables both get_by/2 and all/2. We have a stub for pulling the dn out
of the search terms to replace with a uid present clause, and then
change the base and the scope, which should re-enable get/2
functionality.