Fix Repo.get functionality
Prior to this commit, the `Repo.get` functionality temporarily stopped
working as part of an incremental improvement to handle `Repo.get_by/2`
and `Repo.all/2` methods.

This commit provides the handling necessary to inspect the filter
criteria and update with the distinguished name if required.
With this change, the `Repo.get` functionality is now available
for use.