reinstate get/2 with more robust solution
Prior to this commit, attempting to call Repo.get(Model, dn) would fail
to return any results because, by default, we were attempting to search
the attributes of LDAP object for a dn attribute with the supplied
value; however, LDAP does not store a dn as part of the object or its
attributes and thus we cannot search by attributes to find a dn. Instead
we must shift the base of our search to the supplied dn and remove dn
from the search terms. This commit searches through the search terms
replacing an dn search terms with an objectClass present term, because
it is the most benign and most universal. When we change a dn search
term, we record our new base. We attempt to replace the base, scope, and
filter when we've detected a DN search across a unique DN. Given that
our LDAP library is current incapable of searching across multiple,
specified DN objects, we must error out.