Add Repo.update capability
Prior to this commit, the `ecto_ldap` adapter did not support the
ability to update attributes on an LDAP object.  This commit introduces
the necessary code to enable LDAP object attribute updates via Ecto.
With these changes, consumers of the `ecto_ldap` adapter will be able to
invoke the `Repo.update changeset` command for the purposes of updating
selected attributes on a given LDAP object.