add ldap api sandbox
Prior to this commit the only way to test ecto_ldap was to connect it to
a live ldap server, which is slow, non-portable, and limits automated
testing. This commit adds a sandbox/fake ldap server, which can be used
for both development and automated tests. The sandbox will return
realistic data for some queries (though note that the full search
parameters/filtering has not been built out yet).