Convert sandbox.ex to GenServer
Prior to this commit, calls to the LDAP sandbox would return constant
values for LDAP responses.  With the introduction of update
functionality, it will be necessary to simulate data changes in LDAP.

This commit converts the sandbox to a GenServer for the purposes of
tracking the state of the mock LDAP data in the sandbox.  With these
changes, it will be possible to simulate updates to LDAP and
subsequently return an updated version of the model.