return only the fields specified by select
Prior to this commit, when a user issued a `select` for only a handful
of the fields of the schema, the adapter would return the correct count
of fields, but certainly the incorrect values because we'd take the
attributes in schema order and just take the first X regardless of
whether they were the correct fields or not. This commit inspects the
fields specified by the select and returns only those and in the order
the user specifies.