use Timex master for full ASN.1 GeneralizedTime support
Prior to this commit we assumed a specific variant of the ASN.1
Generalized Time UTC format, which included 0s for fractional seconds
that must exist. Clearly this lead to all sorts of issues. This commit
uses the `{ASN1:GeneralizedTime:Z}` format, which has landed in Timex
master, but not yet a release, to more robustly parse the data returned
from LDAP.