Upgrade to Ecto 2.0
Prior to this commit, the `ecto_ldap` adapter did not have the bindings
or behaviours required by the Ecto 2.0 API.  This commit provides the
functions necessary to conform to the requirements of the Ecto 2.0
adapter behaviours.  With these changes, dependent projects will now
be able to upgrade to Ecto 2.0 without any dependency conflicts from
the `ecto_ldap` project.