Handle list-based select criteria
Prior to this commit, use of the `:select` query criteria resulted in a
failure of the `ecto_ldap` adapter to return results due to a mismatched
function head expectation.

This commit resolves the bug by conforming the expected data to use
flattened lists instead of lists-of-lists, as well as consistent pruning
of the selected attributes to be returned from the adapter.

Tests which exercise the expected response have been added to ensure
proper functioning going forward.