Handle updated Ecto data structures in adapter
Prior to this commit, the adapter expected messages to follow pre-Ecto
2.2 structures and formats.  This resulted in mismatched function heads
and a variety of problems when mapping data into an Ecto model for
consumption by a client.

This commit resolves some of these issues by removing extraneous
functions and leveraging the tranlsation function provided to `execute`
directly.  With the improvements introduced in Ecto 2.2, many of the
data mapping could be removed, making for a simpler way to return data.

This commit does not address the `update` functionality.