Merge Storage behaviour with Adapter
Prior to this commit, the `Ecto.Adapter.Storage` behaviour was specified
in a separate file which could be referenced by consuming applications.
Due to the changes introduced in the API, these behaviours were not
being recognized at runtime.  This change consolidates the behaviour
into the main `Adapter` to resolve this inconsistency.  As a result,
errors and warnings which called out the missing behaviour have been
resolved.